refactor tests
Move all tests that don't validate private state are on integration level.
Reorganise fixtures while at it.
